Ensure all types are correct.
This involves an assert statement at the start of the call to the
objective function.

This is also faster then the current implementation.

-------------------------------------------------------------------------------------- benchmark: 2 tests -------------------------------------------------------------------------------------
Name (time in s)                     Min                Max               Mean            StdDev             Median               IQR            Outliers     OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_optimise (NOW)               9.4730 (1.0)      10.7749 (1.03)      9.8230 (1.0)      0.5371 (46.58)     9.6118 (1.0)      0.3946 (36.35)         1;1  0.1018 (1.0)           5           1
test_optimise (0002_c2d68b1)     10.4035 (1.10)     10.4339 (1.0)      10.4229 (1.06)     0.0115 (1.0)      10.4243 (1.08)     0.0109 (1.0)           1;0  0.0959 (0.94)          5           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

This is faster than the coercion at all calls but only slightly.

-------------------------------------------------------------------------------------- benchmark: 2 tests -------------------------------------------------------------------------------------
Name (time in s)                     Min                Max               Mean            StdDev             Median               IQR            Outliers     OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_optimise (0003_73ed9ac)     10.0485 (1.0)      10.3181 (1.0)      10.1731 (1.0)      0.0964 (1.0)      10.1579 (1.0)      0.0880 (1.0)           2;0  0.0983 (1.0)           5           1
test_optimise (NOW)              10.1513 (1.01)     10.6592 (1.03)     10.3136 (1.01)     0.2251 (2.34)     10.1685 (1.00)     0.3257 (3.70)          1;0  0.0970 (0.99)          5           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

+++ src/optimisation.py
@@ -203,7 +203,7 @@ def create_initial_population(
 
         # add to population
         population.append([primary_allocation, secondary_allocation])
-    return np.array(population)
+    return np.array(population).astype(np.int64)
 
 
 def rank_population(
